+++ reporter.js
@@ -1,94 +1,8 @@
-var yyyymmdd = require('yyyy-mm-dd')
-var stripAnsi = require('strip-ansi')
 var chalk = require('chalk')
-var path = require('path')
-var os = require('os')
+var separators = require('./log-helper.js').separators
+var logHelper = require('./log-helper.js')
 
 var pkg = require('./package.json')
-
-var PADDING = '        '
-var SEPARATOR = os.EOL + os.EOL
-var NEXT_LINE = os.EOL === '\n' ? '\r\v' : os.EOL
-
-function rightPag (str, length) {
-  var add = length - stripAnsi(str).length
-  for (var i = 0; i < add; i++) str += ' '
-  return str
-}
-
-function time (c) {
-  return c.dim('at ' + yyyymmdd.withTime(module.exports.now()))
-}
-
-function line (c, label, color, message) {
-  var labelFormat = c.bold[color].bgBlack.inverse
-  var messageFormat = c.bold[color]
-
-  return rightPag(labelFormat(label), 8) +
-         messageFormat(message) + ' ' +
-         time(c)
-}
-
-function info (c, str) {
-  return line(c, ' INFO ', 'green', str)
-}
-
-function warn (c, str) {
-  return line(c, ' WARN ', 'yellow', str)
-}
-
-function error (c, str) {
-  return line(c, ' ERROR ', 'red', str)
-}
-
-function params (c, type, fields) {
-  var max = 0
-  var current
-  for (var i = 0; i < fields.length; i++) {
-    current = fields[i][0].length + 2
-    if (current > max) max = current
-  }
-  return fields.map(function (field) {
-    return PADDING + rightPag(field[0] + ': ', max) + c.bold(field[1])
-  }).join(NEXT_LINE)
-}
-
-function errorParams (c, type, client) {
-  if (!client) {
-    return ''
-  } else {
-    var user = client.user ? client.user.id : 'unauthenticated'
-    return params(c, 'error', [
-      ['User ID', user],
-      ['Node ID', client.nodeId || 'unknown'],
-      ['Subprotocol', client.sync.otherSubprotocol || 'unknown'],
-      ['IP address', client.remoteAddress]
-    ])
-  }
-}
-
-function note (c, str) {
-  return PADDING + c.grey(str)
-}
-
-function prettyStackTrace (c, err, root) {
-  if (root.slice(-1) !== path.sep) root += path.sep
-
-  return err.stack.split('\n').slice(1).map(function (i) {
-    i = i.replace(/^\s*/, PADDING)
-    var match = i.match(/(\s+at [^(]+ \()([^)]+)\)/)
-    if (!match || match[2].indexOf(root) !== 0) {
-      return c.red(i)
-    } else {
-      match[2] = match[2].slice(root.length)
-      if (match[2].indexOf('node_modules') !== -1) {
-        return c.red(match[1] + match[2] + ')')
-      } else {
-        return c.yellow(match[1] + match[2] + ')')
-      }
-    }
-  }).join(NEXT_LINE)
-}
 
 var reporters = {
 
@@ -104,8 +18,8 @@ var reporters = {
     var dev = app.env === 'development'
 
     return [
-      info(c, 'Logux server is listening'),
-      params(c, 'info', [
+      logHelper.info(c, 'Logux server is listening'),
+      logHelper.params(c, 'info', [
         ['Logux server', pkg.version],
         ['PID', app.options.pid],
         ['Node ID', app.options.nodeId],
@@ -114,21 +28,21 @@ var reporters = {
         ['Supports', app.options.supports],
         ['Listen', url]
       ]),
-      (dev ? note(c, 'Press Ctrl-C to shutdown server') : '')
+      (dev ? logHelper.note(c, 'Press Ctrl-C to shutdown server') : '')
     ]
   },
 
   connect: function connect (c, app, ip) {
     return [
-      info(c, 'Client was connected'),
-      params(c, 'info', [['IP address', ip]])
+      logHelper.info(c, 'Client was connected'),
+      logHelper.params(c, 'info', [['IP address', ip]])
     ]
   },
 
   authenticated: function authenticated (c, app, client) {
     return [
-      info(c, 'User was authenticated'),
-      params(c, 'info', [
+      logHelper.info(c, 'User was authenticated'),
+      logHelper.params(c, 'info', [
         ['User ID', client.user.id],
         ['Node ID', client.nodeId || 'unknown'],
         ['Subprotocol', client.sync.otherSubprotocol],
@@ -141,8 +55,8 @@ var reporters = {
   disconnect: function disconnect (c, app, client) {
     var user = client.user ? client.user.id : 'unauthenticated'
     return [
-      info(c, 'Client was disconnected'),
-      params(c, 'info', [
+      logHelper.info(c, 'Client was disconnected'),
+      logHelper.params(c, 'info', [
         ['User ID', user],
         ['Node ID', client.nodeId || 'unknown'],
         ['IP address', client.remoteAddress]
@@ -152,7 +66,7 @@ var reporters = {
 
   destroy: function destroy (c) {
     return [
-      info(c, 'Shutting down Logux server')
+      logHelper.info(c, 'Shutting down Logux server')
     ]
   },
 
@@ -160,9 +74,9 @@ var reporters = {
     var prefix = err.name + ': ' + err.message
     if (err.name === 'Error') prefix = err.message
     return [
-      error(c, prefix),
-      prettyStackTrace(c, err, app.options.root),
-      errorParams(c, 'error', client)
+      logHelper.error(c, prefix),
+      logHelper.prettyStackTrace(c, err, app.options.root),
+      logHelper.errorParams(c, 'error', client)
     ]
   },
 
@@ -174,15 +88,15 @@ var reporters = {
       prefix = 'SyncError: ' + err.description
     }
     return [
-      error(c, prefix),
-      errorParams(c, 'error', client)
+      logHelper.error(c, prefix),
+      logHelper.errorParams(c, 'error', client)
     ]
   },
 
   clientError: function clientError (c, app, client, err) {
     return [
-      warn(c, 'Client error: ' + err.description),
-      errorParams(c, 'warn', client)
+      logHelper.warn(c, 'Client error: ' + err.description),
+      logHelper.errorParams(c, 'warn', client)
     ]
   }
 
@@ -199,9 +113,5 @@ module.exports = function (type, app) {
 
   return reporter.apply({ }, args).filter(function (i) {
     return i !== ''
-  }).join(NEXT_LINE) + SEPARATOR
-}
-
-module.exports.now = function () {
-  return new Date()
+  }).join(separators.NEXT_LINE) + separators.SEPARATOR
 }
